ACU-545: Refactor Backstop Setup code
