+++ tests/backstop_data/engine_scripts/puppet/activity-feed-panel/activity-detail.js
@@ -7,6 +7,6 @@ module.exports = async (page, scenario, vp) => {
 
   await utility.waitForAngular();
   await utility.waitForLoadingComplete();
-  await page.click('.civicase__activity-feed__list .civicase__activity-card');
+  await page.click('.civicase__activity-feed__list:nth-child(3) .civicase__activity-card');
   await page.waitFor('.blockUI.blockOverlay', { hidden: true });
 };
