+++ tests/backstop_data/gulp-tasks/data-setup-steps/create-sample-upload-file.service.js
@@ -1,0 +1,10 @@
+var fs = require('fs');
+
+/**
+ * Create Sample Upload File
+ */
+function createSampleUploadFile () {
+  fs.writeFileSync('sample.txt', 'Sample Text');
+}
+
+module.exports = createSampleUploadFile;
